Add file
Student class contains information about the student and includes the various classifiers need for sorting.
Sorter class contains the algorithm to sort the student records.
Tester is the main class which reads in student records in a file using the default File class. The student records are then processed via the sorter class and then written to an output file.